@@ -17,47 +17,45 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "@skeletonlabs/skeleton": "3.0.0-next.10",
-    "@skeletonlabs/skeleton-svelte": "1.0.0-next.19",
-    "@sveltejs/adapter-netlify": "^4.4.0",
-    "@sveltejs/adapter-vercel": "^5.5.2",
-    "@sveltejs/kit": "^2.9.0",
-    "@sveltejs/vite-plugin-svelte": "^5.0.1",
-    "@tailwindcss/forms": "^0.5.9",
-    "@tanstack/svelte-query-devtools": "^5.62.3",
+    "@skeletonlabs/skeleton": "3.0.0-next.11",
+    "@skeletonlabs/skeleton-svelte": "1.0.0-next.20",
+    "@sveltejs/adapter-netlify": "^4.4.2",
+    "@sveltejs/adapter-vercel": "^5.6.3",
+    "@sveltejs/kit": "^2.17.3",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tanstack/svelte-query-devtools": "^5.66.11",
     "@types/ms": "^0.7.34",
-    "autoprefixer": "^10.4.20",
     "cross-env": "^7.0.3",
     "dotenv": "^16.4.7",
     "drizzle-kit": "^0.20.18",
-    "jose": "^5.9.6",
+    "jose": "^5.10.0",
     "mode-watcher": "^0.3.1",
-    "postcss": "^8.4.49",
-    "postcss-load-config": "^5.1.0",
     "postgres": "^3.4.5",
-    "prettier": "^3.4.2",
-    "prettier-plugin-svelte": "^3.3.2",
-    "prettier-plugin-tailwindcss": "^0.5.14",
-    "svelte": "^5.9.0",
-    "svelte-check": "^4.1.1",
-    "tailwindcss": "^3.4.16",
+    "prettier": "^3.5.2",
+    "prettier-plugin-svelte": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.11",
+    "sass": "^1.85.1",
+    "svelte": "^5.20.5",
+    "svelte-check": "^4.1.4",
+    "tailwindcss": "^4.0.9",
     "tslib": "^2.8.1",
-    "tsx": "^4.19.2",
-    "typescript": "^5.7.2",
-    "vite": "^6.0.3",
-    "vite-node": "^2.1.8",
-    "vite-plugin-tailwind-purgecss": "^0.3.5"
+    "tsx": "^4.19.3",
+    "typescript": "^5.8.2",
+    "vite": "^6.2.0",
+    "vite-node": "^2.1.9"
   },
   "type": "module",
   "dependencies": {
-    "@floating-ui/dom": "^1.6.12",
-    "@fontsource/roboto": "^5.1.0",
+    "@floating-ui/dom": "^1.6.13",
+    "@fontsource/roboto": "^5.1.1",
     "@neondatabase/serverless": "^0.8.1",
-    "@tanstack/svelte-query": "^5.62.3",
+    "@tailwindcss/vite": "^4.0.9",
+    "@tanstack/svelte-query": "^5.66.11",
     "@trpc/client": "^10.45.2",
     "@trpc/server": "^10.45.2",
     "@upstash/redis": "^1.34.4",
-    "better-auth": "^1.1.18",
+    "better-auth": "^1.1.21",
     "buffer": "^6.0.3",
     "clsx": "^2.1.1",
     "devalue": "^5.1.1",
@@ -65,12 +63,11 @@
     "drizzle-zod": "^0.5.1",
     "lucide-svelte": "^0.395.0",
     "ms": "^2.1.3",
-    "sass": "^1.82.0",
     "superjson": "^2.2.2",
     "svelte-icons-pack": "^3.1.3",
-    "tailwind-merge": "^2.5.5",
+    "tailwind-merge": "^2.6.0",
     "trpc-svelte-query": "^2.1.0",
     "trpc-sveltekit": "^3.6.2",
-    "zod": "^3.23.8"
+    "zod": "^3.24.2"
   }
 }
