@@ -80,7 +80,7 @@
       <button
         type="submit"
         class={cl_join(
-          'btn rounded-md px-2 py-1 text-white outline-none',
+          'btn rounded-md px-2 py-1 text-white outline-hidden',
           !passkey_is_set ? 'bg-green-600' : 'bg-amber-600'
         )}
       >
@@ -120,7 +120,7 @@
         new_item_name_element && new_item_name_element.focus();
       }, 400 + 50);
     }}
-    class="btn gap-1 space-x-1 rounded-lg bg-tertiary-700 px-2 py-1 font-bold text-white outline-none dark:bg-tertiary-700"
+    class="btn bg-tertiary-700 dark:bg-tertiary-700 gap-1 space-x-1 rounded-lg px-2 py-1 font-bold text-white outline-hidden"
   >
     <Icon src={VscAdd} class="text-2xl" />
     <span>Add New</span>
@@ -154,14 +154,14 @@
           }}
         >
           <button
-            class="btn rounded-lg bg-primary-700 px-2 py-1 font-bold text-white dark:bg-primary-700"
+            class="btn bg-primary-700 dark:bg-primary-700 rounded-lg px-2 py-1 font-bold text-white"
           >
             <Icon src={AiOutlinePlus} class="-mx-1 -my-1 text-2xl" />
           </button>
         </ConfirmPopover>
         <button
           onclick={() => (new_item_add_opened = false)}
-          class="btn rounded-md bg-error-600 px-1 py-1 text-white dark:bg-error-500"
+          class="btn bg-error-600 dark:bg-error-500 rounded-md px-1 py-1 text-white"
         >
           <Icon src={RiSystemCloseLargeFill} class="-mt-1 text-xl" />
         </button>
