@@ -1,4 +1,7 @@
 <script lang="ts">
+  import '@fontsource/roboto/latin.css';
+  import '../app.css';
+  import '../app.scss';
   import { type Snippet } from 'svelte';
   import { ModeWatcher } from 'mode-watcher';
   import { QueryClientProvider } from '@tanstack/svelte-query';
@@ -6,15 +9,13 @@
   import { SvelteQueryDevtools } from '@tanstack/svelte-query-devtools';
   import TopAppBar from '~/components/TopAppBar.svelte';
   import CookieCacheRefresh from '../lib/CookieCacheRefresh.svelte';
-  import '@fontsource/roboto/latin.css';
-  import '../app.pcss';
 
   let { children }: { children: Snippet } = $props();
 </script>
 
 <QueryClientProvider client={queryClient}>
   <ModeWatcher />
-  <div class="contaiiner mx-auto mb-1 max-w-screen-lg">
+  <div class="contaiiner mx-auto mb-1 max-w-(--breakpoint-lg)">
     <TopAppBar />
     <div class="mx-2">
       {@render children()}
